+++ src/test.h
@@ -27,6 +27,14 @@ namespace forge
 
   int test_project(const std::filesystem::path& project_directory,
                    const std::optional<std::string>& target,
+                   const std::optional<std::string>& profile,
+                   const std::optional<std::string>& system_profile,
+                   std::span<const std::string_view> arguments,
+                   std::ostream& output,
+                   std::ostream& error);
+
+  int test_project(const std::filesystem::path& project_directory,
+                   const std::optional<std::string>& target,
                    std::span<const std::string_view> arguments,
                    const ProcessRunner& process_runner,
                    std::ostream& output,
@@ -35,6 +43,15 @@ namespace forge
   int test_project(const std::filesystem::path& project_directory,
                    const std::optional<std::string>& target,
                    const std::optional<std::string>& profile,
+                   std::span<const std::string_view> arguments,
+                   const ProcessRunner& process_runner,
+                   std::ostream& output,
+                   std::ostream& error);
+
+  int test_project(const std::filesystem::path& project_directory,
+                   const std::optional<std::string>& target,
+                   const std::optional<std::string>& profile,
+                   const std::optional<std::string>& system_profile,
                    std::span<const std::string_view> arguments,
                    const ProcessRunner& process_runner,
                    std::ostream& output,