+++ src/workspace.h
@@ -55,6 +55,23 @@ namespace forge
   int build_workspace(const std::filesystem::path& workspace_directory,
                       const std::optional<std::string>& project,
                       const std::optional<std::string>& profile,
+                      const std::optional<std::string>& system_profile,
+                      const std::vector<std::string>& compile_definitions,
+                      std::ostream& output,
+                      std::ostream& error);
+
+  int build_workspace(const std::filesystem::path& workspace_directory,
+                      const std::optional<std::string>& project,
+                      const std::optional<std::string>& profile,
+                      const std::vector<std::string>& compile_definitions,
+                      const ProcessRunner& process_runner,
+                      std::ostream& output,
+                      std::ostream& error);
+
+  int build_workspace(const std::filesystem::path& workspace_directory,
+                      const std::optional<std::string>& project,
+                      const std::optional<std::string>& profile,
+                      const std::optional<std::string>& system_profile,
                       const std::vector<std::string>& compile_definitions,
                       const ProcessRunner& process_runner,
                       std::ostream& output,
@@ -90,6 +107,23 @@ namespace forge
                               std::ostream& output,
                               std::ostream& error);
 
+  int build_and_run_workspace(const std::filesystem::path& workspace_directory,
+                              std::string_view selection,
+                              const std::optional<std::string>& profile,
+                              const std::optional<std::string>& system_profile,
+                              std::span<const std::string_view> arguments,
+                              std::ostream& output,
+                              std::ostream& error);
+
+  int build_and_run_workspace(const std::filesystem::path& workspace_directory,
+                              std::string_view selection,
+                              const std::optional<std::string>& profile,
+                              const std::optional<std::string>& system_profile,
+                              std::span<const std::string_view> arguments,
+                              const ProcessRunner& process_runner,
+                              std::ostream& output,
+                              std::ostream& error);
+
   int test_workspace(const std::filesystem::path& workspace_directory,
                      const std::optional<std::string>& selection,
                      const std::optional<std::string>& profile,
@@ -100,6 +134,23 @@ namespace forge
   int test_workspace(const std::filesystem::path& workspace_directory,
                      const std::optional<std::string>& selection,
                      const std::optional<std::string>& profile,
+                     std::span<const std::string_view> arguments,
+                     const ProcessRunner& process_runner,
+                     std::ostream& output,
+                     std::ostream& error);
+
+  int test_workspace(const std::filesystem::path& workspace_directory,
+                     const std::optional<std::string>& selection,
+                     const std::optional<std::string>& profile,
+                     const std::optional<std::string>& system_profile,
+                     std::span<const std::string_view> arguments,
+                     std::ostream& output,
+                     std::ostream& error);
+
+  int test_workspace(const std::filesystem::path& workspace_directory,
+                     const std::optional<std::string>& selection,
+                     const std::optional<std::string>& profile,
+                     const std::optional<std::string>& system_profile,
                      std::span<const std::string_view> arguments,
                      const ProcessRunner& process_runner,
                      std::ostream& output,